* node.h (struct RNode): Change argc from int to long.  Otherwize   NEW_CFUNC() sets argc to a wrong value on platforms where   sizeof(int) != sizeof(long) and the byte order is big-endian.   This fixes breakage on FreeBSD/sparc64.
git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@3263 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ node.h
@@ -140,7 +140,7 @@ typedef struct RNode {
     union {
 	struct RNode *node;
 	ID id;
-	int argc;
+	long argc;
 	VALUE value;
     } u2;
     union {
